requirements.txt
streamlit>=1.24.0
llama-cpp-python>=0.2.0
python-dotenv>=1.0.0
numpy>=1.22.0
gguf>=0.1.0
sentencepiece>=0.1.97
tqdm>=4.65.0
rich>=13.3.0